+++ src/operator/pickData.ts
@@ -1,0 +1,5 @@
+import { map, pipe } from "rxjs";
+
+export function pickData() {
+  return pipe(map<{ data?: any }, any>((x) => x?.data));
+}
