+++ src/operator/toDesc.ts
@@ -1,0 +1,8 @@
+import { catchError, map, of, pipe } from "rxjs";
+
+export function toDesc() {
+  return pipe(
+    map((desc) => ({ data: desc, success: true })),
+    catchError(() => of({ data: undefined, success: true }))
+  );
+}
